scripts/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
PyMuPDF>=1.23.0
Pillow>=10.0.0